Stable/version2.1.2

+++ src/components/Auth/DeleteButton.vue
@@ -1,19 +1,20 @@
 <!-- LogoutButton.vue -->
 <template>
-    <button @click="deleteAllData">Delete user and all data!</button>
+    <button class="bg-red-500 text-white p-3 rounded" @click="deleteAllData">DELETE USER AND ALL DATA!</button>
 </template>
   
 <script>
 import AuthServiceVue from '../../services/AuthService.vue';
+import LocalStorageService from '../../services/LocalStorageService.vue';
 
 export default {
     methods: {
         async deleteAllData() {
             const _this = this
-            if(confirm("Are you sure you want to delete user ? All data wil be deleted")) {
+            if(confirm("Are you sure you want to delete user ? All data and your user wil be deleted")) {
                 AuthServiceVue.deleteUser() // Adjust the URL as needed
                 .then(() => {
-                    localStorage.clear()
+                    LocalStorageService.clear()
                     _this.$router.push({ path: '/' })
                 })
                 .catch(error => {
